adaptacao dos codigos

+++ Mergesort/java/mergeFloatCrescente.java
@@ -1,19 +1,19 @@
 public class mergeFloatCrescente {
 	/*
-	* Descricao: apartir de um vetor de inteiros, essa 
+	* Descricao: apartir de um vetor de reais, essa 
 	* funcao subdivide o vetor em outros dois e intercala 
 	* seus elementos entre si no vetor principal
 	* 
-	* Parametros: um vetor de inteiros (vetor a ser intercalado) e 
+	* Parametros: um vetor de reais (vetor a ser intercalado) e 
 	* 3 inteiros (o primeiro, o ultimo e o indice do elemento do 
 	* meio do vetor)
 	*/
-	public static void intercalar (int[] vet, int esq, int meio, int dir) {
-		int nEsq = (meio -esq) + 1;
+	public static void intercalar (float[] vet, int esq, int meio, int dir) {
+		int nEsq = (meio - esq) + 1;
 		int nDir = dir - meio;
 
-		int[] arrayEsq = new int[nEsq+1];
-		int[] arrayDir = new int[nDir+1];
+		float[] arrayEsq = new float[nEsq+1];
+		float[] arrayDir = new float[nDir+1];
 
 		//Sentinela no final dos dois arrays
 		arrayEsq[nEsq] = 0x7FFFFFFF;
@@ -44,13 +44,13 @@ public class mergeFloatCrescente {
 	}
 
 	/* 
-	* Descricao: essa funcao ordena um vetor de inteiros 
+	* Descricao: essa funcao ordena um vetor de reais 
 	* em ordem crescente com o metodo mergesort
 	* 
-	* Parametro: um vetor de inteiros (vetor a ser ordenado)
+	* Parametro: um vetor de reais (vetor a ser ordenado)
 	* e dois inteiros (indice do primeiro e do ultimo elemento do vetor)
 	*/
-	public static void merge (int[] vet, int esq, int dir) {
+	public static void merge (float[] vet, int esq, int dir) {
 		if (esq < dir) {
 			int meio = (esq + dir) /2;
 			merge(vet, esq, meio);
@@ -60,13 +60,13 @@ public class mergeFloatCrescente {
 	}
 
 	/*
-		* Descricao: essa funcao chama pela funcao merge 
-		* para ordena o vetor em ordem crescente, possuindo 
-		* argumentos mais simplificados
-		* 
-		* Parametro: um vetor de inteiros (vetor a ser ordenado)
-		*/
-	public static void mergesort (int[] vet) {
+	* Descricao: essa funcao chama pela funcao merge 
+	* para ordena o vetor em ordem crescente, possuindo 
+	* argumentos mais simplificados
+	* 
+	* Parametro: um vetor de reais (vetor a ser ordenado)
+	*/
+	public static void mergesort (float[] vet) {
 		merge (vet, 0, vet.length-1);
 	}
